Create new SelectableTextContainer
